requirements.txt
numpy
pandas
setuptools
SQLAlchemy
psycopg2
python-dotenv
scikit-learn